Separate styled component css into separate file
