+++ src/components/CocktailList/CocktailList.js
@@ -1,27 +1,5 @@
 import React from "react";
-import styled from "styled-components/macro";
-
-const Wrapper = styled.ul`
-  color: pink;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(2, 1fr);
-  display: grid;
-  gap: 40px;
-  list-style-type: none;
-  padding-left: 0;
-`;
-
-const Cocktail = styled.li`
- 
-`;
-
-const CocktailImage = styled.img`
-  width: 100%;
-  height: 100%;
-  aspect-ratio: 1/1;
-  object-fit: cover;
-  border: 1px solid #CA0000;
-`;
+import { Wrapper, Cocktail, CocktailImage } from './CocktailList.styled';
 
 
 const CocktailList = () => {
